dataset and md updated

+++ dataset/smartbugs/mycontract.sol
@@ -8,17 +8,15 @@
 pragma solidity ^0.4.24;
 
 contract MyContract {
-
     address owner;
 
     function MyContract() public {
         owner = msg.sender;
     }
 
-    function sendTo(address receiver, uint amount) public {
+    function sendTo(address receiver, uint256 amount) public {
         // <yes> <report> ACCESS_CONTROL
         require(tx.origin == owner);
         receiver.transfer(amount);
     }
-
 }
